Some modifications to avoid a few warnings.

+++ src/abstract/ARKey.cpp
@@ -282,18 +282,18 @@ void ARKey::getKeyArray(std::string inString)
 		// if correct notename, store accidental  at correct position in  mkarray
 		switch(gd_noteName2pc(notename.c_str())) 
 		{
-			case NOTE_C: index=0; fAccarray[0]=acc; break;
+			case NOTE_C: index=0; fAccarray[0]=(float)acc; break;
 			case NOTE_CIS: break; // do nothing (yet)
-			case NOTE_D: index=1; fAccarray[1]=acc; break;
+			case NOTE_D: index=1; fAccarray[1]=(float)acc; break;
 			case NOTE_DIS: break; // do nothing (yet)
-			case NOTE_E: index=2; fAccarray[2]=acc; break;
-			case NOTE_F: index=3; fAccarray[3]=acc; break;
+			case NOTE_E: index=2; fAccarray[2]=(float)acc; break;
+			case NOTE_F: index=3; fAccarray[3]=(float)acc; break;
 			case NOTE_FIS: break; // do nothing (yet)
-			case NOTE_G: index=4; fAccarray[4]=acc; break;
+			case NOTE_G: index=4; fAccarray[4]=(float)acc; break;
 			case NOTE_GIS: break; // do nothing (yet)
-			case NOTE_A: index=5; fAccarray[5]=acc; break;
+			case NOTE_A: index=5; fAccarray[5]=(float)acc; break;
 			case NOTE_AIS: break; // do nothing (yet)
-			case NOTE_H: index=6; fAccarray[6]=acc; break;
+			case NOTE_H: index=6; fAccarray[6]=(float)acc; break;
 			default: return; // if neede, do error handling here
 		}
